ConsumptieRecyclerAdapter klasse om recyclerAdapter code gescheiden te houden
